src/gateway/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
grpcio==1.51.1
grpcio-tools==1.51.1
h11==0.14.0
idna==3.4
protobuf==4.21.12
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
